Refactor: create package "clients"
move classes: "KiwixWebChromeClient" and "KiwixWebViewClents" into the clients package
