+++ app/src/main/java/org/kiwix/kiwixmobile/base/BasePresenter.java
@@ -17,6 +17,9 @@
  */
 package org.kiwix.kiwixmobile.base;
 
+import org.kiwix.kiwixmobile.base.contract.Presenter;
+import org.kiwix.kiwixmobile.base.contract.ViewCallback;
+
 /**
  * Created by EladKeyshawn on 05/04/2017.
  */
