+++ app/src/main/java/org/kiwix/kiwixmobile/base/contract/ViewCallback.java
@@ -15,14 +15,10 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.kiwix.kiwixmobile.base;
+package org.kiwix.kiwixmobile.base.contract;
 
 /**
  * Created by EladKeyshawn on 05/04/2017.
  */
-public interface Presenter<V extends ViewCallback> {
-
-  void attachView(V mvpView);
-
-  void detachView();
+public interface ViewCallback {
 }
