+++ app/src/main/java/org/kiwix/kiwixmobile/views/web/KiwixWebView.java
@@ -35,8 +35,8 @@ import android.webkit.WebView;
 import android.widget.Toast;
 
 import org.kiwix.kiwixmobile.KiwixApplication;
-import org.kiwix.kiwixmobile.KiwixWebChromeClient;
-import org.kiwix.kiwixmobile.KiwixWebViewClient;
+import org.kiwix.kiwixmobile.clients.KiwixWebChromeClient;
+import org.kiwix.kiwixmobile.clients.KiwixWebViewClient;
 import org.kiwix.kiwixmobile.R;
 import org.kiwix.kiwixmobile.utils.LanguageUtils;
 import org.kiwix.kiwixmobile.utils.SharedPreferenceUtil;
