+++ app/src/main/java/org/kiwix/kiwixmobile/zim_manager/fileselect_view/contract/ZimFileSelectViewCallback.java
@@ -17,7 +17,7 @@
  */
 package org.kiwix.kiwixmobile.zim_manager.fileselect_view.contract;
 
-import org.kiwix.kiwixmobile.base.ViewCallback;
+import org.kiwix.kiwixmobile.base.contract.ViewCallback;
 import org.kiwix.kiwixmobile.library.entity.LibraryNetworkEntity;
 
 import java.util.ArrayList;
